+++ src/pages/index.jsx
@@ -1,6 +1,6 @@
 import Footer from "../components/footer";
 import Navbar from "../components/navbar";
-import HeroHeader from "../components/hero-header";
+import HeroHeader from "../components/homepage-landing";
 
 export default function Home() {
   return (
